Backend: implementando o Canonical Data Model Pattern no BackEnd
