+++ backend/src/controllers/ClassroomController.js
@@ -1,11 +1,11 @@
-const classroomModel = require('../models/ClassroomModel');
+const classroomRepo = require('../repositories/ClassroomRepository');
 const express = require('express');
 const router = express.Router();
 const { generateHash } = require('../util/hash');
 
 const getAllClassrooms = router.get('/teacher/classroom', async (request, response) => {
     try {
-        const classrooms = await classroomModel.getAllClassrooms();
+        const classrooms = await classroomRepo.getAllClassrooms();
         return response.status(200).json(classrooms);
     } catch (error) {
         console.error('Error getting all classrooms:', error);
@@ -17,7 +17,7 @@ const getAllClassroomByTeacherId = router.get('/teacher/classroom/:id', async (r
     try {
         const teacherid = request.params.id;
         
-        const classrooms = await classroomModel.getAllClassroomByTeacherId(teacherid);
+        const classrooms = await classroomRepo.getAllClassroomByTeacherId(teacherid);
         
         if (!classrooms.length) {
             return response.status(404).json({ message: "Teacher not found or no classrooms assigned" });
@@ -34,7 +34,7 @@ const createClassroom = router.post('/teacher/classroom/create', async (request,
     try {
         const { classroomname, classroomdescription, teacherid, trackid, moduleid } = request.body;
         tokenclass = generateHash(Date.now());
-        const newClassroom = await classroomModel.createClassroom({classroomname, classroomdescription, teacherid, trackid, moduleid, tokenclass});
+        const newClassroom = await classroomRepo.createClassroom({classroomname, classroomdescription, teacherid, trackid, moduleid, tokenclass});
         return response.status(201).json({ message: "Classroom created successfully", newClassroom });
     } catch (error) {
         console.error('Error creating classroom:', error);
@@ -47,7 +47,7 @@ const updateClassroom = router.put('/teacher/classroom/update', async (request, 
         const classroomid = request.params.id;
         const { classroomname, classroomdescription, teacherid, trackid, moduleid } = request.body;
         const tokenclass = generateHash();
-        const updatedClassroom = await classroomModel.updateClassroom({classroomid, classroomname, classroomdescription, teacherid, trackid, moduleid, tokenclass});
+        const updatedClassroom = await classroomRepo.updateClassroom({classroomid, classroomname, classroomdescription, teacherid, trackid, moduleid, tokenclass});
         return response.status(200).json({ message: "Classroom updated successfully", updatedClassroom });
     } catch(error){
         return response.status(500).json({ message: error.message ||  "Internal server error" });
